requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.11
django-allauth==0.41.0
django-betterforms==1.2
django-nested-formset==0.1.4
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
